@@ -1,56 +1,58 @@
 {
   "name": "@digicatapult/wasp-ws-reading-service",
-  "version": "0.0.8",
+  "version": "0.1.0",
   "description": "Service for WASP",
   "main": "app/index.js",
   "scripts": {
-    "test": "NODE_ENV=test mocha --config ./test/mocharc.js ./test",
+    "test": "NODE_ENV=test mocha --config ./test/mocharc.json ./test",
     "lint": "eslint .",
     "depcheck": "depcheck",
     "start": "node app/index.js",
     "dev": "NODE_ENV=dev nodemon app/index.js | pino-colada",
-    "coverage": "nyc npm run test"
+    "coverage": "c8 npm run test"
   },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/digicatapult/wasp-ws-reading-service.git"
   },
   "engines": {
-    "node": ">=14.16.0"
+    "node": ">=18.13.0"
   },
   "keywords": [
     "WASP"
   ],
   "author": "Digital Catapult",
   "license": "Apache-2.0",
+  "type": "module",
   "bugs": {
     "url": "https://github.com/digicatapult/wasp-ws-reading-service/issues"
   },
   "homepage": "https://github.com/digicatapult/wasp-ws-reading-service#readme",
   "dependencies": {
-    "@asyncapi/parser": "^1.9.1",
-    "dotenv": "^8.2.0",
-    "envalid": "^7.0.0",
-    "express": "^4.17.1",
+    "@asyncapi/parser": "^1.17.2",
+    "dotenv": "^16.0.3",
+    "envalid": "^7.3.1",
+    "express": "^4.18.2",
     "express-ws": "^5.0.2",
-    "kafkajs": "^1.15.0",
-    "pino": "^6.11.1",
-    "pino-http": "^5.5.0"
+    "kafkajs": "^2.2.3",
+    "pino": "^8.8.0",
+    "pino-http": "^8.3.1"
   },
   "devDependencies": {
-    "chai": "^4.3.1",
-    "chai-json": "^1.0.0",
+    "@babel/eslint-parser": "^7.19.1",
+    "@babel/plugin-syntax-import-assertions": "^7.20.0",
+    "c8": "^7.12.0",
+    "chai": "^4.3.7",
     "delay": "^5.0.0",
-    "depcheck": "^1.4.0",
-    "eslint": "^7.21.0",
-    "eslint-config-prettier": "^8.1.0",
-    "eslint-plugin-prettier": "^3.3.1",
-    "mocha": "^8.3.0",
-    "nodemon": "^2.0.7",
-    "nyc": "^15.1.0",
-    "pino-colada": "^2.1.0",
-    "prettier": "^2.2.1",
-    "supertest": "^6.1.3",
-    "ws": "^8.2.2"
+    "depcheck": "^1.4.3",
+    "eslint": "^8.32.0",
+    "eslint-config-prettier": "^8.6.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "mocha": "^10.2.0",
+    "nodemon": "^2.0.20",
+    "pino-colada": "^2.2.2",
+    "prettier": "^2.8.3",
+    "supertest": "^6.3.3",
+    "ws": "^8.12.0"
   }
 }
