@@ -1,8 +1,10 @@
-const envalid = require('envalid')
-const dotenv = require('dotenv')
+import envalid from 'envalid'
+import dotenv from 'dotenv'
 
 if (process.env.NODE_ENV === 'test') {
   dotenv.config({ path: 'test/test.env' })
+} else {
+  dotenv.config()
 }
 
 const vars = envalid.cleanEnv(
@@ -11,7 +13,6 @@ const vars = envalid.cleanEnv(
     SERVICE_TYPE: envalid.str({ default: 'wasp-ws-reading-service'.toUpperCase().replace(/-/g, '_') }),
     LOG_LEVEL: envalid.str({ default: 'info', devDefault: 'debug' }),
     PORT: envalid.port({ default: 80, devDefault: 3000 }),
-    API_MAJOR_VERSION: envalid.str({ default: 'v1' }),
 
     KAFKA_LOG_LEVEL: envalid.str({
       default: 'nothing',
@@ -21,7 +22,7 @@ const vars = envalid.cleanEnv(
       const kafkaSet = new Set(input === '' ? [] : input.split(','))
       if (kafkaSet.size === 0) throw new Error('At least one kafka broker must be configured')
       return [...kafkaSet]
-    })({ default: 'localhost:9092' }),
+    })({ default: ['localhost:9092'] }),
     KAFKA_READINGS_NOTIFICATIONS_TOPIC: envalid.str({ default: 'reading-notifications' }),
     WS_PING_INTERVAL_MS: envalid.num({ default: 500 }),
   },
@@ -30,4 +31,4 @@ const vars = envalid.cleanEnv(
   }
 )
 
-module.exports = vars
+export default vars
