@@ -1,15 +1,19 @@
-const { Kafka, logLevel: kafkaLogLevels } = require('kafkajs')
-const logger = require('./logger')
-const { KAFKA_BROKERS, KAFKA_LOG_LEVEL, KAFKA_READINGS_NOTIFICATIONS_TOPIC } = require('./env')
+import { Kafka, logLevel as kafkaLogLevels } from 'kafkajs'
+import logger from './logger.js'
+import env from './env.js'
+
+const { KAFKA_BROKERS, KAFKA_LOG_LEVEL, KAFKA_READINGS_NOTIFICATIONS_TOPIC } = env
 
 const setupReadingsConsumer = async ({ forwardParams }) => {
   const kafkaLogger = logger.child({ module: 'kafkajs-readings', level: 'error' })
-  const logCreator = () => ({ label, log }) => {
-    const { message } = log
-    kafkaLogger[label.toLowerCase()]({
-      message,
-    })
-  }
+  const logCreator =
+    () =>
+    ({ label, log }) => {
+      const { message } = log
+      kafkaLogger[label.toLowerCase()]({
+        message,
+      })
+    }
 
   const kafka = new Kafka({
     clientId: 'reading-streaming-service',
@@ -51,4 +55,4 @@ const setupReadingsConsumer = async ({ forwardParams }) => {
   }
 }
 
-module.exports = setupReadingsConsumer
+export default setupReadingsConsumer
